test: move workspace variants build test under pixi_build
The test_workspace_variants_separate_work_directories test resolved its
pixi-build-python backend from the channel because it lived in the
top-level tests/integration_python/ directory, outside the autouse
setup_build_backend_override fixture in pixi_build/conftest.py. After the
pixi-build-api-version bump to 5 (#6328) no published backend satisfies
the new version, so the channel solve fails. Moving the test under
pixi_build/ applies the local-backend override on every platform.
